Bugfix: Last part of long messages will no longer be cut off
Problem was, that the loop before only accounted in indices which started before the chunksize of 1999.

So if there are 2500 characters are remaining everything worked and it sent about 1800 of them. But the remaining 700 would not be sent since the while condition would no longer be true.

The new solution runs till all characters are sent. And if there happen to be only 700 characters remaining it just sends those off to discord without any issue.